fix carousel card background
