+++ src/screens/home/Home.tsx
@@ -3,7 +3,7 @@ import About from "./about";
 import HeroSection from "./HeroSection";
 import HowTo from "./how-to";
 import Partner from "./partner";
-import Features from "../features/Features";
+import Features from "./features/Features";
 import CarouselCompany from "../company-carousel/CompanyCarousel";
 
 const Home = (): ReactElement => {
