+++ src/screens/home/features/Features.tsx
@@ -62,7 +62,7 @@ const Features = () => {
     setOpen((open) => !open);
   };
   return (
-    <section className="flex h-fit w-full flex-col place-items-center bg-[url('@/assets/bgCarousel.png')] bg-cover bg-no-repeat pt-20 pb-32">
+    <section className="flex h-fit w-full flex-col place-items-center bg-[url('@/assets/bgCarousel.png')] bg-cover bg-no-repeat pb-32 pt-20">
       <h1 className="mb-14 text-center font-mono text-5xl font-semibold text-white lg:text-[96px]">
         PRODUCTS
       </h1>
@@ -75,6 +75,18 @@ const Features = () => {
             >
               <Card className="rounded-[2.5rem] border-x-[1.5px] border-y-2 border-x-green-600 border-y-green-500 bg-black/25 shadow-inner shadow-green-700">
                 <CardContent className="flex h-[439px] w-[358px] flex-col place-content-center items-center space-y-5 rounded-[2.5rem] border-0 p-0">
+                  <div
+                    className="inherit absolute -z-50 flex h-[439px] w-[358px] rounded-[2.5rem] border-[#414141] bg-[#414141]/50 bg-black backdrop-blur-xl lg:basis-1/3"
+                    style={{
+                      clipPath: " polygon(100% 0%, 0% 0%, 0% 63%, 100% 23%)",
+                    }}
+                  ></div>
+                  <div
+                    className="inherit absolute -z-50 h-[439px] w-[358px] rounded-[2.5rem] bg-green-400/20 backdrop-blur-xl"
+                    style={{
+                      clipPath: "polygon(100% 100%, 0% 100%, 0% 60%, 100% 20%)",
+                    }}
+                  ></div>
                   <img
                     src={data.image}
                     className="h-[50%] w-auto object-contain"
@@ -84,7 +96,7 @@ const Features = () => {
                     {data.name}
                   </h1>
                   <Button
-                    className="w-20 rounded-3xl bg-green-500 text-black"
+                    className="w-20 rounded-3xl bg-green-500 text-black hover:bg-green-400"
                     onClick={(e) => handleOpen(e, data)}
                   >
                     More
